Use better practice for key prop in Alert example
From React docs, the "key" prop in a list of components should be "a string that uniquely identifies a list item among its siblings". Using the item index as a key should be a last resort in case we don't have such identifiers, but here we can use the variant name.